+++ packages/core/createReusableTemplate/index.test.ts
@@ -80,4 +80,21 @@ describe.skipIf(isVue2)('createReusableTemplate', () => {
 
     expect(wrapper.text()).toBe('GoodbyeHi')
   })
+
+  it('hyphen props', () => {
+    const [DefineFoo, ReuseFoo] = createReusableTemplate<{ myMsg: string }>()
+
+    const wrapper = mount({
+      render() {
+        return h(Fragment, null, [
+          h(DefineFoo, ({ $slots, ...args }: any) => h('pre', JSON.stringify(args))),
+          h(ReuseFoo, { myMsg: 'Foo' }),
+          // @ts-expect-error Vue automatically converts hyphenized props to camelCase
+          h(ReuseFoo, { 'my-msg': 'Bar' }),
+        ])
+      },
+    })
+
+    expect(wrapper.text()).toBe('{"myMsg":"Foo"}{"myMsg":"Bar"}')
+  })
 })
